Minor test fix

+++ src/lib/util/MakeTempDirectory.h
@@ -7,7 +7,8 @@
 class MakeTempDirectory
 {
 public:
-	MakeTempDirectory()
+	MakeTempDirectory(bool cleanup=true)
+		: _cleanup(cleanup)
 	{
 		_path = std::tmpnam(nullptr);
 		std::experimental::filesystem::create_directory(_path);
@@ -15,8 +16,11 @@ public:
 
 	~MakeTempDirectory()
 	{
-		std::error_code ec;
-		std::experimental::filesystem::remove_all(_path, ec);
+		if (_cleanup)
+		{
+			std::error_code ec;
+			std::experimental::filesystem::remove_all(_path, ec);
+		}
 	}
 
 	std::experimental::filesystem::path path() const
@@ -25,5 +29,6 @@ public:
 	}
 
 protected:
+	bool _cleanup;
 	std::experimental::filesystem::path _path;
 };
